+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.9.0] - 2025-08-12
+
 - utils/VideoRecorder : check filename extension (must be ".mp4")
 - core/RenderContext : add setSwapchainParameters()
 - multibody : enable displaying forces applies to frame (with lifetime) (https://github.com/Simple-Robotics/candlewick/pull/93)
@@ -244,7 +246,8 @@ This is the first release to use a changelog.
 - Visual bug (uninitialized data) when disabling SSAO
 
 
-[Unreleased]: https://github.com/Simple-Robotics/candlewick/compare/v0.8.0...HEAD
+[Unreleased]: https://github.com/Simple-Robotics/candlewick/compare/v0.9.0...HEAD
+[0.9.0]: https://github.com/Simple-Robotics/candlewick/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/Simple-Robotics/candlewick/compare/v0.8.0...v0.8.0
 [0.8.0]: https://github.com/Simple-Robotics/candlewick/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/Simple-Robotics/candlewick/compare/v0.6.0...v0.7.0